Delete duplicate create reef function
