+++ packages/api/src/reef-applications/reef-applications.controller.ts
@@ -3,14 +3,12 @@ import {
   Body,
   Param,
   Get,
-  Post,
   Put,
   UseInterceptors,
   ClassSerializerInterceptor,
   SerializeOptions,
   Query,
   NotFoundException,
-  Req,
   UseGuards,
 } from '@nestjs/common';
 import { ReefApplicationsService } from './reef-applications.service';
@@ -21,12 +19,7 @@ import {
 } from './dto/update-reef-application.dto';
 import { idFromHash, isValidId } from '../utils/urls';
 import { Auth } from '../auth/auth.decorator';
-import {
-  CreateReefApplicationDto,
-  CreateReefWithApplicationDto,
-} from './dto/create-reef-application.dto';
 import { Public } from '../auth/public.decorator';
-import { AuthRequest } from '../auth/auth.types';
 import { IsReefAdminGuard } from '../auth/is-reef-admin.guard';
 import { ParseHashedIdPipe } from '../pipes/parse-hashed-id.pipe';
 import { OverrideLevelAccess } from '../auth/override-level-access.decorator';
@@ -40,19 +33,6 @@ import { AdminLevel } from '../users/users.entity';
 })
 export class ReefApplicationsController {
   constructor(private reefApplicationsService: ReefApplicationsService) {}
-
-  @Post()
-  async create(
-    @Req() request: AuthRequest,
-    @Body('reefApplication') reefApplication: CreateReefApplicationDto,
-    @Body('reef') reef: CreateReefWithApplicationDto,
-  ): Promise<ReefApplication> {
-    return this.reefApplicationsService.create(
-      reefApplication,
-      reef,
-      request.user,
-    );
-  }
 
   @OverrideLevelAccess(AdminLevel.SuperAdmin, AdminLevel.ReefManager)
   @UseGuards(IsReefAdminGuard)
