+++ packages/api/src/reef-applications/reef-applications.module.ts
@@ -5,15 +5,10 @@ import { ReefApplicationsService } from './reef-applications.service';
 import { EntityExists } from '../validations/entity-exists.constraint';
 import { Reef } from '../reefs/reefs.entity';
 import { ReefApplication } from './reef-applications.entity';
-import { Region } from '../regions/regions.entity';
 import { AuthModule } from '../auth/auth.module';
-import { User } from '../users/users.entity';
 
 @Module({
-  imports: [
-    AuthModule,
-    TypeOrmModule.forFeature([Reef, ReefApplication, Region, User]),
-  ],
+  imports: [AuthModule, TypeOrmModule.forFeature([Reef, ReefApplication])],
   controllers: [ReefApplicationsController],
   providers: [ReefApplicationsService, EntityExists],
 })
